+++ services/horizon/internal/db2/history/offers_test.go
@@ -41,6 +41,18 @@ var (
 		Flags:  2,
 		Amount: xdr.Int64(500),
 	}
+	threeEurOffer = xdr.OfferEntry{
+		SellerId: twoEurOfferSeller,
+		OfferId:  xdr.Int64(50),
+		Buying:   eurAsset,
+		Selling:  nativeAsset,
+		Price: xdr.Price{
+			N: 3,
+			D: 1,
+		},
+		Flags:  2,
+		Amount: xdr.Int64(500),
+	}
 )
 
 func assertOfferEntryMatchesDBOffer(t *testing.T, offerEntry xdr.OfferEntry, offer Offer, lastModifiedLedger xdr.Uint32) {
@@ -109,12 +121,22 @@ func assertOfferEntryMatchesDBOffer(t *testing.T, offerEntry xdr.OfferEntry, off
 	}
 }
 
+func insertOffer(q *Q, offer xdr.OfferEntry, lastModifiedSeq uint32) error {
+	batch := q.NewOffersBatchInsertBuilder(0)
+	err := batch.Add(offer, xdr.Uint32(lastModifiedSeq))
+	if err != nil {
+		return err
+	}
+	return batch.Exec()
+}
+
 func TestGetOfferByID(t *testing.T) {
-	tt := test.Start(t).Scenario("base")
+	tt := test.Start(t)
 	defer tt.Finish()
+	test.ResetHorizonDB(t, tt.HorizonDB)
 	q := &Q{tt.HorizonSession()}
 
-	_, err := q.InsertOffer(eurOffer, 1234)
+	err := insertOffer(q, eurOffer, 1234)
 	tt.Assert.NoError(err)
 	offer, err := q.GetOfferByID(int64(eurOffer.OfferId))
 	tt.Assert.NoError(err)
@@ -122,8 +144,9 @@ func TestGetOfferByID(t *testing.T) {
 }
 
 func TestGetNonExistentOfferByID(t *testing.T) {
-	tt := test.Start(t).Scenario("base")
+	tt := test.Start(t)
 	defer tt.Finish()
+	test.ResetHorizonDB(t, tt.HorizonDB)
 	q := &Q{tt.HorizonSession()}
 
 	_, err := q.GetOfferByID(12345)
@@ -131,23 +154,40 @@ func TestGetNonExistentOfferByID(t *testing.T) {
 }
 
 func TestQueryEmptyOffers(t *testing.T) {
-	tt := test.Start(t).Scenario("base")
+	tt := test.Start(t)
 	defer tt.Finish()
+	test.ResetHorizonDB(t, tt.HorizonDB)
 	q := &Q{tt.HorizonSession()}
 
 	offers, err := q.GetAllOffers()
 	tt.Assert.NoError(err)
 	tt.Assert.Len(offers, 0)
+
+	updated, err := q.GetUpdatedOffers(0)
+	tt.Assert.NoError(err)
+	tt.Assert.Len(updated, 0)
+
+	count, err := q.CountOffers()
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(0, count)
+
+	numRemoved, err := q.CompactOffers(100)
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(int64(0), numRemoved)
+	seq, err := q.GetOfferCompactionSequence()
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(uint32(100), seq)
 }
 
 func TestInsertOffers(t *testing.T) {
-	tt := test.Start(t).Scenario("base")
+	tt := test.Start(t)
 	defer tt.Finish()
+	test.ResetHorizonDB(t, tt.HorizonDB)
 	q := &Q{tt.HorizonSession()}
 
-	_, err := q.InsertOffer(eurOffer, 1234)
+	err := insertOffer(q, eurOffer, 1234)
 	tt.Assert.NoError(err)
-	_, err = q.InsertOffer(twoEurOffer, 1235)
+	err = insertOffer(q, twoEurOffer, 1235)
 	tt.Assert.NoError(err)
 
 	offers, err := q.GetAllOffers()
@@ -161,27 +201,58 @@ func TestInsertOffers(t *testing.T) {
 
 	assertOfferEntryMatchesDBOffer(t, eurOffer, offersByID[eurOffer.OfferId], 1234)
 	assertOfferEntryMatchesDBOffer(t, twoEurOffer, offersByID[twoEurOffer.OfferId], 1235)
+
+	count, err := q.CountOffers()
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(2, count)
+
+	numRemoved, err := q.CompactOffers(12350)
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(int64(0), numRemoved)
+	seq, err := q.GetOfferCompactionSequence()
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(uint32(12350), seq)
+
+	afterCompactionCount, err := q.CountOffers()
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(2, afterCompactionCount)
+
+	afterCompactionOffers, err := q.GetAllOffers()
+	tt.Assert.NoError(err)
+	tt.Assert.Len(afterCompactionOffers, 2)
 }
 
 func TestUpdateOffer(t *testing.T) {
-	tt := test.Start(t).Scenario("base")
+	tt := test.Start(t)
 	defer tt.Finish()
+	test.ResetHorizonDB(t, tt.HorizonDB)
 	q := &Q{tt.HorizonSession()}
 
-	rowsAffected, err := q.InsertOffer(eurOffer, 1234)
+	err := insertOffer(q, eurOffer, 1234)
 	tt.Assert.NoError(err)
-	tt.Assert.Equal(int64(1), rowsAffected)
 
 	offers, err := q.GetAllOffers()
 	tt.Assert.NoError(err)
 	tt.Assert.Len(offers, 1)
+
+	updatedOffers, err := q.GetUpdatedOffers(1233)
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(offers, updatedOffers)
+
+	updatedOffers, err = q.GetUpdatedOffers(100)
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(offers, updatedOffers)
+
+	updatedOffers, err = q.GetUpdatedOffers(1234)
+	tt.Assert.NoError(err)
+	tt.Assert.Len(updatedOffers, 0)
 
 	assertOfferEntryMatchesDBOffer(t, eurOffer, offers[0], 1234)
 
 	modifiedEurOffer := eurOffer
 	modifiedEurOffer.Amount -= 10
 
-	rowsAffected, err = q.UpdateOffer(modifiedEurOffer, 1235)
+	rowsAffected, err := q.UpdateOffer(modifiedEurOffer, 1235)
 	tt.Assert.NoError(err)
 	tt.Assert.Equal(int64(1), rowsAffected)
 
@@ -189,48 +260,100 @@ func TestUpdateOffer(t *testing.T) {
 	tt.Assert.NoError(err)
 	tt.Assert.Len(offers, 1)
 
+	updatedOffers, err = q.GetUpdatedOffers(1234)
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(offers, updatedOffers)
+
+	updatedOffers, err = q.GetUpdatedOffers(1235)
+	tt.Assert.NoError(err)
+	tt.Assert.Len(updatedOffers, 0)
+
 	assertOfferEntryMatchesDBOffer(t, modifiedEurOffer, offers[0], 1235)
 }
 
 func TestRemoveNonExistantOffer(t *testing.T) {
-	tt := test.Start(t).Scenario("base")
+	tt := test.Start(t)
 	defer tt.Finish()
+	test.ResetHorizonDB(t, tt.HorizonDB)
 	q := &Q{tt.HorizonSession()}
 
-	_, err := q.RemoveOffer(xdr.Int64(12345))
+	numAffected, err := q.RemoveOffer(xdr.Int64(12345), 1236)
 	tt.Assert.NoError(err)
+	tt.Assert.Equal(int64(0), numAffected)
 }
 
 func TestRemoveOffer(t *testing.T) {
-	tt := test.Start(t).Scenario("base")
+	tt := test.Start(t)
 	defer tt.Finish()
+	test.ResetHorizonDB(t, tt.HorizonDB)
 	q := &Q{tt.HorizonSession()}
 
-	_, err := q.InsertOffer(eurOffer, 1234)
+	err := insertOffer(q, eurOffer, 1234)
 	tt.Assert.NoError(err)
 	offers, err := q.GetAllOffers()
 	tt.Assert.NoError(err)
 	tt.Assert.Len(offers, 1)
 	assertOfferEntryMatchesDBOffer(t, eurOffer, offers[0], 1234)
 
-	rowsAffected, err := q.RemoveOffer(eurOffer.OfferId)
+	expectedUpdates := offers
+	rowsAffected, err := q.RemoveOffer(eurOffer.OfferId, 1236)
 	tt.Assert.Equal(int64(1), rowsAffected)
 	tt.Assert.NoError(err)
+	expectedUpdates[0].LastModifiedLedger = 1236
+	expectedUpdates[0].Deleted = true
 
 	offers, err = q.GetAllOffers()
 	tt.Assert.NoError(err)
 	tt.Assert.Len(offers, 0)
+
+	offers, err = q.GetOffersByIDs([]int64{int64(expectedUpdates[0].OfferID)})
+	tt.Assert.NoError(err)
+	tt.Assert.Len(offers, 0)
+
+	_, err = q.GetOfferByID(int64(expectedUpdates[0].OfferID))
+	tt.Assert.True(q.NoRows(err))
+
+	updated, err := q.GetUpdatedOffers(1234)
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(expectedUpdates, updated)
+
+	count, err := q.CompactOffers(1235)
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(int64(0), count)
+
+	updated, err = q.GetUpdatedOffers(1234)
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(expectedUpdates, updated)
+
+	count, err = q.CompactOffers(1236)
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(int64(1), count)
+	seq, err := q.GetOfferCompactionSequence()
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(uint32(1236), seq)
+
+	updated, err = q.GetUpdatedOffers(1234)
+	tt.Assert.NoError(err)
+	tt.Assert.Len(updated, 0)
 }
 
 func TestGetOffers(t *testing.T) {
-	tt := test.Start(t).Scenario("base")
+	tt := test.Start(t)
 	defer tt.Finish()
+	test.ResetHorizonDB(t, tt.HorizonDB)
 	q := &Q{tt.HorizonSession()}
 
-	_, err := q.InsertOffer(eurOffer, 1234)
+	err := insertOffer(q, eurOffer, 1234)
 	tt.Assert.NoError(err)
-	_, err = q.InsertOffer(twoEurOffer, 1235)
+	err = insertOffer(q, twoEurOffer, 1235)
 	tt.Assert.NoError(err)
+
+	// check removed offers aren't included in GetOffer queries
+	err = insertOffer(q, threeEurOffer, 1234)
+	tt.Assert.NoError(err)
+	count, err := q.RemoveOffer(threeEurOffer.OfferId, 1235)
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(int64(1), count)
 
 	pageQuery, err := db2.NewPageQuery("", false, "", 10)
 	tt.Assert.NoError(err)