+++ services/horizon/internal/expingest/processors/offers_processor.go
@@ -8,15 +8,20 @@ import (
 	"github.com/stellar/go/xdr"
 )
 
+// The offers processor can be configured to trim the offers table
+// by removing all offer rows which were marked for deletion at least 100 ledgers ago
+const offerCompactionWindow = uint32(100)
+
 type OffersProcessor struct {
-	offersQ history.QOffers
+	offersQ  history.QOffers
+	sequence uint32
 
 	cache *io.LedgerEntryChangeCache
 	batch history.OffersBatchInsertBuilder
 }
 
-func NewOffersProcessor(offersQ history.QOffers) *OffersProcessor {
-	p := &OffersProcessor{offersQ: offersQ}
+func NewOffersProcessor(offersQ history.QOffers, sequence uint32) *OffersProcessor {
+	p := &OffersProcessor{offersQ: offersQ, sequence: sequence}
 	p.reset()
 	return p
 }
@@ -31,14 +36,12 @@ func (p *OffersProcessor) ProcessChange(change io.Change) error {
 		return nil
 	}
 
-	err := p.cache.AddChange(change)
-	if err != nil {
+	if err := p.cache.AddChange(change); err != nil {
 		return errors.Wrap(err, "error adding to ledgerCache")
 	}
 
 	if p.cache.Size() > maxBatchSize {
-		err = p.Commit()
-		if err != nil {
+		if err := p.flushCache(); err != nil {
 			return errors.Wrap(err, "error in Commit")
 		}
 		p.reset()
@@ -47,7 +50,7 @@ func (p *OffersProcessor) ProcessChange(change io.Change) error {
 	return nil
 }
 
-func (p *OffersProcessor) Commit() error {
+func (p *OffersProcessor) flushCache() error {
 	changes := p.cache.GetChanges()
 	for _, change := range changes {
 		var rowsAffected int64
@@ -69,7 +72,7 @@ func (p *OffersProcessor) Commit() error {
 			action = "removing"
 			offer := change.Pre.Data.MustOffer()
 			offerID = offer.OfferId
-			rowsAffected, err = p.offersQ.RemoveOffer(offer.OfferId)
+			rowsAffected, err = p.offersQ.RemoveOffer(offer.OfferId, p.sequence)
 		default:
 			// Updated
 			action = "updating"
@@ -95,6 +98,22 @@ func (p *OffersProcessor) Commit() error {
 	err := p.batch.Exec()
 	if err != nil {
 		return errors.Wrap(err, "error executing batch")
+	}
+	return nil
+}
+
+func (p *OffersProcessor) Commit() error {
+	if err := p.flushCache(); err != nil {
+		return errors.Wrap(err, "error flushing cache")
+	}
+
+	if p.sequence > offerCompactionWindow {
+		// trim offers table by removing offers which were deleted before the cutoff ledger
+		if offerRowsRemoved, err := p.offersQ.CompactOffers(p.sequence - offerCompactionWindow); err != nil {
+			return errors.Wrap(err, "could not compact offers")
+		} else {
+			log.WithField("offer_rows_removed", offerRowsRemoved).Info("Trimmed offers table")
+		}
 	}
 
 	return nil