+++ services/horizon/CHANGELOG.md
@@ -11,6 +11,7 @@ file. This project adheres to [Semantic Versioning](http://semver.org/).
 * Use the Horizon DB instead of Horizon's in memory order book graph to query orderbook details for the /order_book endpoint.
 * Remove JSON variant of `GET /metrics`, both in the server and client code. It's using Prometheus format by default now.
 * Decreased a memory usage of initial state ingestion stage and state verifier ([#2618](https://github.com/stellar/go/pull/2618)).
+* Remove `--exp-ingest-in-memory-only` Horizon flag. The in memory order book graph which powers the path finding endpoints is now updated using the Horizon DB instead of directly via ingestion ([#2630](https://github.com/stellar/go/pull/2630)).
 
 ## v1.3.0